\name{dates}
\alias{dates}
\alias{dates.fts}
\title{ Extract Dates }
\description{
  Extract dates from an Fts object
}
\usage{
dates(x)
}
\arguments{
  \item{x}{ An Fts object }
}
\details{
  extracts the dates attribute from an Fts object
}
\value{
  a vector of dates
}
\author{ Whit Armstrong }
\examples{
x <- fts()
dates(x)
}
\keyword{ts}

